[AdminBundle] Remove unused RoleInterface on the group entity
